apps/api/requirements.txt
# Agno Multi-Agent Platform - Requirements
# Python 3.11+

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async
asyncio>=3.4.3
aiofiles>=23.2.1
httpx>=0.26.0

# AI/LLM
agno>=1.0.0
openai>=1.10.0
anthropic>=0.18.0
groq>=0.4.0

# Database & Storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
redis>=5.0.0
pgvector>=0.2.0

# Memory & RAG
chromadb>=0.4.22
sentence-transformers>=2.3.0
langchain>=0.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pyyaml>=6.0
orjson>=3.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# WebSocket
websockets>=12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
